+++ externals/kyuubi-jdbc-engine/src/main/scala/org/apache/kyuubi/engine/jdbc/deploy/JdbcYarnModeSubmitter.scala
@@ -49,29 +49,12 @@ object JdbcYarnModeSubmitter extends EngineYarnModeSubmitter {
   override def engineMainClass(): String = JdbcSQLEngine.getClass.getName
 
   /**
-   * Jar list for the Hive engine.
+   * Jar list for the JDBC engine.
    */
   override def engineExtraJars(): Seq[File] = {
-    val hadoopCp = sys.env.get("HIVE_HADOOP_CLASSPATH")
     val extraCp = kyuubiConf.get(ENGINE_JDBC_EXTRA_CLASSPATH)
     val jars = new ListBuffer[File]
-    hadoopCp.foreach(cp => parseClasspath(cp, jars))
     extraCp.foreach(cp => parseClasspath(cp, jars))
     jars.toSeq
-  }
-
-  private[jdbc] def parseClasspath(classpath: String, jars: ListBuffer[File]): Unit = {
-    classpath.split(":").filter(_.nonEmpty).foreach { cp =>
-      if (cp.endsWith("/*")) {
-        val dir = cp.substring(0, cp.length - 2)
-        new File(dir) match {
-          case f if f.isDirectory =>
-            f.listFiles().filter(_.getName.endsWith(".jar")).foreach(jars += _)
-          case _ =>
-        }
-      } else {
-        jars += new File(cp)
-      }
-    }
   }
 }
